pyproject.toml
[project]
name = "mediguard-clinical-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.135.2",
    "python-jose>=3.5.0",
    "argon2-cffi>=23.1.0",
    "pycryptodome>=3.20.0",
    "aiokafka>=0.12.0,<0.13.0",
    "redis>=5.0.1",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.1",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.29",
    "psycopg2-binary>=2.9.11",
    "uvicorn>=0.42.0",
    "uuid6>=2025.0.1",
    "zstandard>=0.25.0",
    "faust-streaming[rocksdb]>=0.11.0",
]
